+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.4] - 2018-01-18
+### Added
+- Media component. (Still needs the new measurement component)
+
+### Changed
+- Item component now uses measurement type to determen if it should send it to /mes/:measurementId or /med/:mediaId
+- Router path changed from "/media/:mediaId" to "/med/:mediaId"
+
 ## [0.0.3] - 2018-01-18
 ### Fixed
 - Spelling mistake in README (line 7)